Início da alteração no projeto para atender as demandas de câmera e integração com pagamento
