+++ readme.md
@@ -1,2 +1,5 @@
 Criação do Aplicativo em Expo com:
 - Inclusão da tela de câmera e configuração geral para execução do aplicativo
+
+17/05/2024:
+- Leitura de Materiais Relacionados com Pagamentos e início da organização do projeto para atender as demandas de câmera e integrações com pagamento
